requirements.txt
openai
azure-functions
azure-storage-blob